+++ devicemanagement/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   of that management service.
 - [SW-13218] Re-enabling the getModuleMemorySizeMB and getModuleMemoryVendorPartNumber tests for silicon.
 - `getModulePCIENumPortsMaxSpeed` test renamed to `getModulePCIEPortsMaxSpeed`.
+- Disabled getModulePCIEPortsMaxSpeed test on SysEMU.
 ### Deprecated
 ### Removed
 - Workaround for SP Trace